Update test results using tools/run_hjk_tests.sh
There are many changes since it's been a while since the
script was last run. Didn't carefully check the changes
(we have the better test-mk.bash based testing method now)
but it seems that most of the changes are due to a new formatting
of the test results.

(This script should be kept around mostly for testing the demo lexicon.)